# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quick Turn",
        "description": "Documentación de API QuickTurn",
        "contact": {
            "email": "germanca8@correo.ugr.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/turnos": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Lista todas los turnos",
                "description": "Endpoint para listar los turnos",
                "operationId": "6adc8ab3972fa2e0a40a9ddec9b865b9",
                "responses": {
                    "200": {
                        "description": "Lista todas los turnos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Turno"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Crea un turno",
                "description": "Endpoint para almacenar un turno",
                "operationId": "99bc216a5722c88def2faab3dfae2a13",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreTurno"
                },
                "responses": {
                    "200": {
                        "description": "Turno creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "444": {
                        "$ref": "#/components/responses/444"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/turnos/{turnoID}": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Muestra un turno por ID",
                "description": "Endpoint para mostrar un turno",
                "operationId": "699035c27931dc669dac2f892f00e894",
                "parameters": [
                    {
                        "name": "turnoID",
                        "in": "path",
                        "description": "ID del turno a obtener",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del turno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Elimina un turno",
                "description": "Endpoint para eliminar un turno",
                "operationId": "b555e1b000489a87c4e7bf6d8472add3",
                "parameters": [
                    {
                        "name": "turnoID",
                        "in": "path",
                        "description": "ID del turno a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Actualiza un turno",
                "description": "Endpoint para actualizar un turno",
                "operationId": "facac0d2429ef5eb6263c29342001480",
                "parameters": [
                    {
                        "name": "turnoID",
                        "in": "path",
                        "description": "ID del turno a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTurno"
                },
                "responses": {
                    "200": {
                        "description": "Turno actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/turnos/{seccionID}": {
            "post": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Lista todos los turnos de 1 seccion",
                "description": "Endpoint para mostrar todas los turnos asociadas a una seccion",
                "operationId": "acebe29969101c30f7c5a8501fc43432",
                "parameters": [
                    {
                        "name": "seccionID",
                        "in": "path",
                        "description": "ID de la sección",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista todos los turnos de 1 seccion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/turnos/{seccionID}/eliminar": {
            "delete": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Elimina todos los turnos de una sección",
                "description": "Endpoint para eliminar un turno",
                "operationId": "492ab0d1708b1f1c6853542f70bfc588",
                "parameters": [
                    {
                        "name": "seccionID",
                        "in": "path",
                        "description": "ID de la seccion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turnos eliminados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Turno": {
                "properties": {
                    "id": {
                        "description": "Id asociado",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "seccion_id": {
                        "description": "Id de la seccion",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "invitado_id": {
                        "description": "Id del invitado",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "numTurno": {
                        "description": "Numero de turno",
                        "type": "integer",
                        "maxLength": 255,
                        "example": "1"
                    },
                    "fechaTurno": {
                        "description": "Fecha y hora de solicitud de turno",
                        "type": "string",
                        "maxLength": 52,
                        "example": "2022-03-15T13:23:02.000000Z"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última modificación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "404": {
                "description": "No query results for model",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "444": {
                "description": "Name its already used",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Name is already used"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "requestBodies": {
...
            "StoreTurno": {
                "description": "Turno que necesita ser creado",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "seccion_id",
                                "invitado_id",
                                "numTurno",
                                "fechaTurno"
                            ],
                            "properties": {
                                "seccion_id": {
                                    "$ref": "#/components/schemas/Turno/properties/seccion_id"
                                },
                                "invitado_id": {
                                    "$ref": "#/components/schemas/Turno/properties/invitado_id"
                                },
                                "numTurno": {
                                    "$ref": "#/components/schemas/Turno/properties/numTurno"
                                },
                                "fechaTurno": {
                                    "$ref": "#/components/schemas/Turno/properties/fechaTurno"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateTurno": {
                "description": "Turno que necesita ser actualizado",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "numTurno",
                                "fechaTurno"
                            ],
                            "properties": {
                                "numTurno": {
                                    "$ref": "#/components/schemas/Turno/properties/numTurno"
                                },
                                "fechaTurno": {
                                    "$ref": "#/components/schemas/Turno/properties/fechaTurno"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Token Auth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}